Centralize Bungie API key and request builders
