+++ src/scripts/oauth/oauth.service.js
@@ -1,4 +1,5 @@
 import angular from 'angular';
+import { bungieApiUpdate } from '../services/bungie-api-utils';
 
 export function OAuthService($q, $injector, localStorageService, OAuthTokenService) {
   'ngInject';
@@ -14,27 +15,12 @@ export function OAuthService($q, $injector, localStorageService, OAuthTokenServi
   function refreshToken() {
     const $http = $injector.get('$http');
 
-    var apiKey;
-    if ($DIM_FLAVOR === 'release' || $DIM_FLAVOR === 'beta') {
-      if (window.chrome && window.chrome.extension) {
-        apiKey = $DIM_API_KEY;
-      } else {
-        apiKey = $DIM_WEB_API_KEY;
-      }
-    } else {
-      apiKey = localStorageService.get('apiKey');
-    }
-
-    return $http({
-      method: 'POST',
-      url: 'https://www.bungie.net/Platform/App/GetAccessTokensFromRefreshToken/',
-      headers: {
-        'X-API-Key': apiKey
-      },
-      data: {
+    return $http(bungieApiUpdate(
+      '/Platform/App/GetAccessTokensFromRefreshToken/',
+      {
         refreshToken: OAuthTokenService.getRefreshToken().value
       }
-    })
+    ))
     .then((response) => {
       if (response && response.data && (response.data.ErrorCode === 1) && response.data.Response && response.data.Response.accessToken) {
         const inception = Date.now();
@@ -54,15 +40,10 @@ export function OAuthService($q, $injector, localStorageService, OAuthTokenServi
     });
   }
 
-  function revokeToken() {
-    // Revokes token via api
-  }
-
   return {
     isAuthenticated,
     getToken,
-    refreshToken,
-    revokeToken
+    refreshToken
   };
 }
